Update enums.ts, .spec.ts, and 32 more files...
