+++ src/components/LayerTree/LayerTree.spec.tsx
@@ -1,11 +1,11 @@
-import { render, screen,  } from '@testing-library/react';
-import { describe, it } from 'vitest';
+import { render, screen } from '@testing-library/react';
+import { describe, it, expect } from 'vitest';
 
 import LayerTree from './LayerTree';
-import { MapViewerBase } from '../MapViewer/MapViewerContext';
-import {karteRpFeatures as testdata} from '../../OwsContext/tests/data'
+import { OwsContextBase } from '../../react-ows-lib/ContextProvider/OwsContextBase';
+import {karteRpFeatures as testdata} from '../../ows-lib/OwsContext/tests/data'
 
-import {expect} from '@testing-library/jest-dom'
+
 
 import {ReactNode} from 'react'
 
@@ -15,19 +15,18 @@ const getKarteRpFeatures = () => {
 
 const MapViewerBaseWrapper = ({ children }: { children: ReactNode }) => {
   return (
-    <MapViewerBase initialFeatures={getKarteRpFeatures()}>
+    <OwsContextBase initialFeatures={getKarteRpFeatures()}>
       {children}
-    </MapViewerBase>
+    </OwsContextBase>
   )
 }
 
-
-describe('Dragging test', () => {
-  it('Drag Node works as expected', () => {
+describe('LayerTree', () => {
+  it('LayerTree renders with initial data', () => {
 
     render(<LayerTree />, {wrapper: MapViewerBaseWrapper});
     
-    expect(screen.getByLabelText('Karte RP')).toBeInTheDocument();
+    expect(screen.getByText('Karte RP')).toBeInTheDocument();
 
   });
 });
