+++ src/components/SecurityEditor/Forms.tsx
@@ -1,6 +1,6 @@
 import { TextInput, useCreateContext, SimpleForm, CreateBase, useInput, type TextInputProps } from 'react-admin'
 import { useEffect, type ReactNode, useState } from 'react'
-import { useMapViewerContext } from '../MapViewer/MapViewerContext'
+import { useOwsContextBase } from '../../react-ows-lib/ContextProvider/OwsContextBase'
 import SchemaAutocompleteArrayInput from '../../jsonapi/components/SchemaAutocompleteArrayInput'
 import SchemaAutocompleteInput from '../../jsonapi/components/SchemaAutocompleteInput'
 
@@ -8,7 +8,7 @@ export const GeoJSONInput = ({
   source,
   ...rest
 }: TextInputProps): ReactNode => {
-  const { geoJSON, setGeoJSON } = useMapViewerContext()
+  const { geoJSON, setGeoJSON } = useOwsContextBase()
   const { field: { onChange } } = useInput({ source })
 
   const [isOver, setIsOver] = useState(false)
