+++ src/ows-lib/XMLParser/parseCapabilities.ts
