+++ src/ows-lib/XMLParser/tests/capabilitites_130.xml
