+++ src/ows-lib/XMLParser/types.ts
